[JAX] Allow registering callbacks to be called when backends are cleared
`jax.extend.backend` allows the user to register a callback that will be called
when JAX backends are cleared. The primary purpose is to let the user clear any
caches that hold a reference to JAX backends transitively (via JAX
`Sharding`/`Mesh`/`Device`) so that it can help destroy cleared backends.

PiperOrigin-RevId: 768173340