[pallas] Fix shard_map + Megacore in TPU interpret mode.
The fix involves threading the axis sizes and device ID through the
interpretation code, so that we only query the axis sizes and axis indices
once at the start of a kernel on a device.

Also fixes a bug in the TPU interpret mode validation code that runs at the
end of each kernel invocation.