[JAX] Move the fallback of `colocated_cpu_devices` logic from the colocated Python test to the API
The colocated Python test defines a fallback logic for `colocated_cpu_devices`, which would find CPU devices from a local CPU backend instead of the default backend. This works for typical setups of McJAX, and is intended to be temporary until McJAX backends define its own CPU devices.

This change moves the fallback logic from the test into the colocated
Python API and make the `colocated_cpu_devices` API invoke it when necessary.
The fallback logic was inaccessible from the API, which
made it difficult to write another high-level API for colocated Python  such
as `colocated_cpu_mesh` (which will follow subsequently) and have it tested
on McJAX.

PiperOrigin-RevId: 769854488