Add hermetic `nvshmem` dependencies to JAX targets.
JAX wheel build rule implementation is also updated to exclude accidental dependencies on NVSHMEM libraries in the wheel content. If the wheel needs to be built with these dependencies, provide `--@local_config_nvshmem//:override_include_nvshmem_libs=True` in Bazel options.

NVSHMEM binaries are included in the dependencies if CUDA binary dependencies are added as well, e.g. `--@local_config_cuda//:enable_cuda`.

NVSHMEM libraries are included in the dependencies if `--@local_config_nvshmem//:include_nvshmem_libs=True` (the default flag value is `False`). Please note that this is a temporary solution, and it should be removed after GLIBC is updated on RBE runners. At the moment `libnvshmem.so` files can't be linked to the targets because they are built with GLIBC version higher than on RBE runners. In the future `--@local_config_cuda//cuda:include_cuda_libs=True` should be used.

PiperOrigin-RevId: 770663391