[JAX] Relax the return type of `colocated_python` decorator
`colocated_python` decorator wraps a function, and the returned function has a
special method `specialize` that lets the user provide explicit information of
the output spec or execution devices. This `specialize` method is in principle
not a part of `Callable` protocol, so access to it would not be valid typing.

This change relaxes the return type of `colocated_python` to `Any` so that
`specialize` method access does not cause typing check failure.

PiperOrigin-RevId: 772206576