[JAX] Add a cache around HLO lowering rules.
The MLIR Python bindings are slow, relative to building MLIR from C++. This change updates the MLIR lowering to build each primitive in an out-of-line function the first time we see it. Then, both that time and subsequent times, we can inline that function to stamp out a copy of the rule.

PiperOrigin-RevId: 772600579