[Mosaic:TPU][Relayouts] Express some retilings as unpack+pack and remove offset restrictions
The (packing, 128) -> (1, packing * 128) and (8, 128) -> (packing * 8, 128) are just unpack+pack

This factors out the body of ext_op_rule_impl and trunc_op_rule_impl into vregPack and vregUnpack. It also removes the op parameter from getPackUnpackSpec since we don't have one in relayout - it was a weird parameter to have anyway, but may hurt error reporting.

PiperOrigin-RevId: 772791112