[Mosaic GPU][NFC] Make FragmentedArray.warp_dim into a sequence
This allows us to partition the 4 warps over 2 different dimensions. So far, we've
never had the need to do it, but it will be necessary to support some of the weirder
TMEM layouts (e.g. 2CTA with M=64, which splits both rows and columns over warps).

This change doesn't use any of the new features and simply changes how we represent
the layouts. It's arguably even a simplification, since we rarely treat lane_dims
(that were a sequence) from warp_dims (that was a single dim).

PiperOrigin-RevId: 773624014