Deprecate jax.dtypes.SUPPORTED_DTYPES and add is_supported_dtype()
The current set confusingly contains jax ScalarMeta objects instead of actual dtypes
This means for example that jnp.float32 in SUPPORTED_DTYPES returns True, but
np.float32 in SUPPORTED_DTYPES returns False. Because dtypes are so commonly duck-typed,
it's better for us to provide an API that handles dtype conversions automatically.