Move `jax._src.cudnn.*` sources to their own build rule
As part of this, I opted to import some APIs into the top level of `jax._src.numpy`, so that internal submodules that depend on it don't have to manage individual source files. This required a bit of refactorization of `jax.numpy` sources and imports, as well as renaming of `jax/_src/numpy/array.py` to avoid shadowing the `array` function.

PiperOrigin-RevId: 775362445