[jaxpr consts] Propagate closed-over constants to the top-level jit (take 2)
We want to move all closed-over constants from `pjit_p` primitives
to the enclosing scope, similarly to what we do for `scan_p`, and
`cond_p`.

We don't want to change the behavior for the top-level `jit` because that would
require changing the calling convention for executables. Thus, we make
an exception for the top-level jit, and we let it contain a
`ClosedJaxpr` with the closed-over constants in the whole
program.

See #29679 for background.

We add a flag `--jax_use_simplified_jaxpr_constants` to enable this
change. This flag is set to `False` for now.