[jaxpr consts] Propagate closed-over constants to the top-level jit
We want to move all closed-over constants from `pjit_p` primitives
to the enclosing scope, similarly to what we do for `scan_p`, and
`cond_p`.

The top-level `pjit_p` will contain all the closed-over constants in the program.
We don't change the behavior for the top-level `jit` because that would
require changing the calling convention for executables.

See #29679 for background.