[Mosaic] Extend indirect DMA verification to dynamic shapes
Verify static dimension sizes, but ignore dynamic dimension sizes -- the latter can be verified at runtime.

The main difference is that before shapes `<a, b>` and `<a, ?>` (where `a` and `b` are statically known) would be treated as non-equal, and now `b` and `?` will not be compared leaving it to the lowering pass to insert a dynamic check.

Note that compared to `tpu.enqueue_dma`, for `tpu.enqueue_indirect_dma` we are not enforcing that the number of dynamic dimensions in source and target must match.

PiperOrigin-RevId: 775877232