[mutable-arrays] fix autodiff with scan
The scan transpose rule put all cotangents corresponding to consts into the
carry in the transposed scan, and passed in UndefinedPrimals for the primal env
in backward_pass. But when differentiating with respect to refs, we need the
cotangent refs to be in the primal env (according to our current convention),
and we need to keep them in the scan consts.

So this PR rearranges things accordingly.

Co-authored-by: Dougal Maclaurin <dougalm@google.com>