[Pallas:MGPU] Provide output layout hint if a primitive is followed by a layout cast
This lets us avoid the pattern of adding a plgpu specific primitive each time
we need to provide layout hints to the lane lowering. With the new approach,
plgpu.load and plgpu.broadcasted_iota are simple functions that apply the lax
operation and immediately follow it by a layout cast. This allows us to avoid
duplicating lowering rules just to add support for more flexible layout hints.

PiperOrigin-RevId: 776463753