[Mosaic GPU] Add a new API to help with copying matmul scales to TMEM
This surfaces the `tcgen05.cp` instruction, that can asynchronously transfer
data from SMEM to TMEM. The function is currently specialized to transferring
scales for block-scaled matmuls, which is why it expects such unconventional
shapes (the layouts are honestly a bit mind bending).

Logically, the scales e.g. for the left hand side should be thought of as a
(128, K // 32) array (for 8-bit matmul). However, they need to be transformed
by first reshaping to (4, 32, K // 32) and then transposing to (32, 4, K // 32)
before the kernel begins. This format satisfies the layout requirements of
the TensorCore... Unfortunately the data is too fine grained for us to be able
to fix it up during the TMA transfer.

This is why the function expects (32, 4, K // 32) SMEM references, which are physically,
but not logically, contiguous. The TMEM references are (128, K // 32),
corresponding to the logical view of the data (and their layout being responsible
for representing the weirdness that hides underneath).

PiperOrigin-RevId: 779128725