Clarify `wait_recv` semantics
I'm new to distributed Pallas kernels and for a long time was confused when reading the documentation, incorrectly assuming `wait_recv` means waiting until the target device has received the data. In fact I believe it means waiting until the current device has received the data (which is presumably sent from yet another third device that is neither the current device nor the target device in the all-gather example).