[Mosaic GPU] Implement `meet` and `join` for replicated layouts.
We no longer assume that two layouts within a `{Least,Most}ReplicatedExpression`
are always compatible, and now make sure to reduce all the layout expressions
within the replicated expression before deciding whether a least/most
replicated expression even exists.

This is the correct behaviour to have for equations. The logic implemented
previously is appropriate for `Hint`s, and will be added to the layout
inference in an upcoming change.

PiperOrigin-RevId: 780487035