[Mosaic GPU] Allow the loose extraction of assignments from hints.
By "loose", we mean here that we perform an extraction of a layout assuming
that all the layouts that have not been resolved yet in the expression are
compatible with all the other layouts.

This is not a safe simplification, since nothing prevents us from constructing
replicated expressions where some layouts are incompatible with each other.
Nevertheless, when consuming hints, we're OK with temporarily introducing an
assignment that may lead to an unsatisfiable system---since we will simply
backtrack and pick another assignment in that case.

PiperOrigin-RevId: 780520998