[Pallas TPU] Fix bug where we wouldn't pass in a mesh context during lowering because we'd fail to detect comms in the Jaxpr
When lowering collectives (e.g. semaphore_signal, async_remote_copy), Pallas provides a "mesh context" to the lowering that allows the primitive to convert the device_id from a MESH type (i.e. tuple of coordinates) to a LOGICAL type (linearized index), using the mesh shape. If this is not provided, Pallas throws a check-fail type error (indicating the missing mesh strides) because there should be no case where we are using a collective but not providing the mesh info to the collective during lowering. This can happen though, because in order to know whether or not there are collectives in the kernel and pass down the mesh context, we look for `NamedAxisEffect`s which come from primitives that use named axes (e.g. `axis_index`). However, if the kernel does not use any of these primitives, this analysis will decide that there is no communication in the kernel and not pass down a mesh context. However, we can (currently) write kernels that do use collectives without any NamedAxisEffects.

For example, in some cases where we do a barrier w/o using the axis name, we'd have comms in the kernel but not explicitly use axis names (see the new test in this CL).

To address this issue, we also create a dummy comms effect that informs the analysis that there is indeed communication and therefore we need the mesh context. In the long run though, all collective primitives will have named axis effects so the original analysis will work again.

PiperOrigin-RevId: 780636959