[Pallas TPU] Add lookahead for input buffers to emit_pipeline.
This option enables arbitrary-depth lookahead where input buffers will fetch as many future blocks as slots are available. In the case of repeated blocks, the lookahead will continue to scan forwards to find the next changed block and cache a pointer to that block so that future lookaheads will start scanning from that position.

PiperOrigin-RevId: 780754452