[Pallas:MGPU] Create a programatic way to derive reduced layouts from tiled layouts in Pallas
Back when we only really had WGMMA layout it was ok to add WGMMA_ROW and WGMMA_COL
as special cases, but with the proliferation of various TCGEN05 layouts this is
no longer feasible. This change adds a `.reduce` method on `plgpu.Layout` that lets you
obtain its reduced form, without us having to do anything.

This change also modifies the lowering of broadcasts so that it tries a reasonable set
of candidate output layouts to find one that could work for the broadcast. It only succeeds
if it finds exactly one result, as we know that WGMMA_COL and TCGEN05_COL are the same and
cannot be effectively disambiguated.

I kept the existing *_ROW and *_COL layouts in the enum for backwards compatibility and
will do a cleanup later. They should be considered deprecated from now on.

PiperOrigin-RevId: 780939955