[Mosaic GPU] Get rid of `func.FuncOp`s in `gpu_layout_inference_test`.
This is not an op that we ever intend to support, and propagating layouts with
regards to it was a bit messy. Instead, we just use `llvm.mlir_undef` to
produce input values directly now.

This forces us to handle it in both the legacy and equational layout inference,
and uncovers a propagation bug in the legacy inference.

PiperOrigin-RevId: 781557307