[Mosaic GPU][NFC] Enable the gmem argument to `async_copy` to be an `ir.BlockArgument`
This is needed in upcoming changes, where we will create a block in the lowering of the `CustomPrimitiveOp`. Kernel arguments are accessed inside the block via the block argument. However, also inside the lowering of `CustomPrimitiveOp` we will call `async_copy` if the body contains it. `async_copy` creates a tma descriptor and for initializing this descriptor on the host, we need to get to the original `unrealized_conversion_cast`.

PiperOrigin-RevId: 781581079