[jaxprs] The next step in ClosedJaxpr simplification
Now that after #29882 non-scalar constants appear as `core.Literal`
in `Jaxpr.eqns.invars`, and `ClosedJaxpr` are mostly empty of `consts`.
Before this change the lowering would serialize the constant as
a HLO constant. There are several problems with this:

  * if the constant is on a device, it gets copied to the host and
  is embedded in the lowered HLO. This increases memory usage,
  and if the constant was sharded on multiple devices it will
  now become replicated.
  * if a constant appears in multiple places in the Jaxprs, today
  we will make copies.
  * if the constants are inlined in the HLO the XLA constant folding
  starts to constant-fold them, and we see warnings of very slow
  constant folding.
  * the constant folding changes the numerical behavior compared
  to current lowering strategy.

With this change we hoist the constants out of the Jaxpr (the
new function `jaxpr_hoisted_consts`, we pass them as argument to
the top-level function and we thread them to other functions
that need them. We add new arguments to functions to represent
the hoisted constants.

This change has no effect, unless JAX_USE_SIMPLIFIED_JAXPR_CONSTANTS=True.

This is still an experiment.