[Mosaic GPU] Add constraints to `EquationSystem`.
Constraints are essentially equations from which no assignment can be derived, and
that instead represent arbitrary assertions we want our solution to validate.

For now, we only add one such constraint: `Relayout`. `Relayout` states that
it must be possible for its `source` layout to be relaid out into its `target`
layout.

In lieu of modelling `Relayout` as a theoretical concept, our constraint models
details of our implementation. This will allow us to make the system
`Unsatisfiable` if theoretically compatible (but undesirable) layouts are
inferred from `Hint`s, and therefore to let layout inference backtrack and
attempt to find a better solution.

PiperOrigin-RevId: 782937649