Fix to allow slicing into replicated dims
I added an assert on slice_sizes to be full, but did not realize offset_dims can also be sliced dims (i.e. are in in start_index_map). We do make the check that all indexed dims are replicated, but the check on slice_size re-did this and raised an error erroneously.

We can either:
- remove the extraneous slice_sizes check
- add an extra check that non-full slices are on replicated operand

I've added the extra check version for extra caution.

PiperOrigin-RevId: 783326135