test_multiprocess: only initialize GPUs that exist
Without this, a multi-process test requesting 2 processes with one GPU
each would crash on a system with only one GPU, because the 2nd process
would attempt to initialize a non-existent 2nd GPU.

With this change, only 1 process will be initialized. Test cases can
decide how to handle a device/process count of 1.

Closes jax-ml/jax#29513.