[Mosaic GPU] Introduce constraints in the equational layout inference.
This allows fixing a quirk in hint derivation, by ensuring that wrong choices
of hints can be backtracked from safely. The data flow of the graph imposes
relayout constraints such that, for a variable `v`

* `Relayout(producer(v), v)`, and
* `forall c : consumers(v), Relayout(v, c)`

must hold. We now derive these constraints automatically.

As a consequence of the change, `test_infer_splat_layout_for_vector_splat`
needed to be modified to either allow propagating `splat`s up towards the `undef`
op (in which case, the `undef` op's output would always be assigned a `splat`
layout), or forbidding `splat` layouts for the `undef` op's output
systematically (in which case, the `undef` op's output would always be
assigned a `strided` layout). We decided to go with the former, and eliminated
the non-splat case from the test.

Similarly `test_infer_layout_from_yield_op_in_layouts_for_for_op` needed to
be modified, since the behaviour no longer matches the one of the old
implementation (but the new behaviour is correct, since `undef` is allowed to
be `splat`).

PiperOrigin-RevId: 784530390