[Mosaic GPU] Fix barrier indexing in tests.
When we started using normal `Barrier`s in lieu of `TMABarrier`s to wait on
`tcgen05.mma` completion, we forgot to replace the barrier passed to
`tcgen05.commit_arrive` with the new MMA barrier.

(Un)luckily, because the barrier was always the last TMA barrier allocated and
since we don't have any bounds check when slicing `BarrierRef`s, the code still
ended up accessing the right barrier under the hood and was therefore still
working.

Added a bounds check to `BarrierRef.__getitem__` to prevent this in the future.

PiperOrigin-RevId: 784680090