Fix numerical bugs in the gradients of sigmoid/logistic, tanh, expm1.
For sigmoid, its derivative was computed as sigmoid'(x) = sigmoid(x) * (1 - sigmoid(x)), which is inaccurate when sigmoid(x) is close to 1, and actually caused the gradient to vanish for x > ~17 even though it should only vanish at ~88.

A more accurate way to compute the gradient would be e.g. sigmoid(x) * sigmoid(-x).

For expm1 the derivative was computed as `expm1'(x) = expm1(x) + 1` which is inaccurate for large negative x, this CL replaces it with `expm1'(x) = exp(x)`.

For tanh the derivative was computed as `tanh'(x) = (1 + tanh(x)) (1 - tanh(x))`, this CL replaces it with `tanh'(x) = 4 * sigmoid(2x) * sigmoid(-2x)` which doesn't suffer from numerical issues when x is large.

Due to possible performance issues stemming from not reusing computations as much, this changes modify only the behavior when accuracy is set to HIGHEST.

PiperOrigin-RevId: 784912803