Separate out `vma` and `unreduced/reduced` in the ShapedArray str_short().
We are doing this because `vma` and `unreduced/reduced` are very different to each other. So it makes sense to split them apart in the repr too.

Before: `f32[1,1,1,1]{V:c, U:a, R:b}`, `f32[1,1,1,1]{V:c, U:a}`

After: `f32[1,1,1,1]{U:a, R:b}{c}`, `f32[1,1,1,1]{U:a}{c}`
PiperOrigin-RevId: 784981854