Correctly apply global scale in nvfp4 quantization.
Currently, the global scale is not applied consistently to the input
tensor, leading to all values being rounded to zeroes in most cases
when using the global scale logic from the test.

The existing tests did not catch this because they use matrix
multiplications on dequantized tensors as the reference. This ended
up just comparing 0@0 to 0@0.

Instead of adjusting the matmul tests, I opted to add test coverage
for quantization logic directly, since that doesn't require carefully
chosen tolerances in most cases.