[jaxpr consts] Fix const args handling in shard_map shardy lowering
The shardy lowering for shard map was also broken for shape polymorphism.
When we enter the block for SdyManualComputation we need to pass
the dim var values and the const arg values to the block as arguments.

Drive-by fixing of a test in api_test
(for JAX_USE_SIMPLIFIED_JAXPR_CONSTANTS=True)