[Mosaic GPU] Add a method for efficient SMEM->GMEM copies without TMA
The copies go through registers for now, but it should be easy to use the
Ampere-style async instructions instead. I verified that we don't create any
bank conflicts and that the global accesses are well coalesced when possible.
More improvements are possible when the swizzle is small, but I left that for
the next change.

PiperOrigin-RevId: 788952504