Initial commit to add memory_space to JAX types i.e memory space in types (MSIT).
**Motivation**

Add memory_spaces to values (ShapedArray) akin to sharding-in-types for host-offloading (data + compute). Why?
  * So that we get better errors at trace time rather than at compile time from XLA which are super hard to debug
  * So that users can inspect/reflect on the memory_spaces during trace time.
  * In activation offloading, we can remove `policy.src` from offloading remat policies since they can be inferred from the type instead!

**Design**

We add `memory_space` as an argument to `ShapedArray` which is an enum containing 2 values; `Device` and `Host`. `Device` means `Hbm | Vmem | Smem | ...`. `Device` is kind of like Auto Mode of sharding-in-types but without the complex machinery.

We choose `Device` instead of more granular memory spaces because XLA can decide intermediates to live in `Vmem` (for example) and JAX should allow that. So annotating arrays with Device encompasses all the device memory spaces that XLA might choose. But `Host` is a bit special because JAX has host offloading support (data + compute host offload).

We also choose `Device` because it's device agnostic. Other memory spaces might be different on different backends (TPU, GPU, etc).

**What are the memory space propagation rules?**

The propagation rules are pretty simple.

* memory_space is propagated to the output if all input memory spaces match. If they don't match, it's an error.
  * Device + Device = Device
  * Device + Host = error
  * Host + Host = Host
* `jax.device_put` is capable of changing the memory_space of output since only that API is used for transfers to/from host and device.

**How does this relate to AbstractRef.memory_space?**

AbstractRefs are used by pallas and jax.ArrayRef and AbstractRef also has a memory_space argument.

But we should have some rules on how AbstractRef.memory_space acts under pallas_call and outside pallas_call.

AbstractRef.memory_space can be more granualar than ShapedArray.memory_space:

* in pallas, can be: Host, HBM, Vmem, Any (=Device?)
* in jax, can be: all those plus Device
* but in jax it's illegal to perform operations other than pallas_call on AbstractRefs unless the memory space is Host or Device, ie:
`def get_abstract_eval_rule(ref, ...): if in_jax(): assert ref.memory_space in {Host, Device}`

But one valid question is why even have Hbm, Vmem on AbstractRef in JAX (outside pallas_call)? It's because jax.ArrayRef can be created on such memory_spaces and passed to pallas_call. In other words, ArrayRef can be passed between pallas_calls and since pallas can represent Hbm, Vmem, Smem, etc so should AbstractRef.

But our abstract_eval rules can be context sensitive on whether we are under pallas_call (or core_map) or not and error out as said above.

This is TBD though and can change in the future.

**Subtyping**

In the future, JAX might want to add more granualar memory_spaces on ShapedArray like Hbm, Vmem, etc. So how will the current system work with that? We can use subtyping for that!

`Hbm`, `Vmem`, etc can be subtypes of `Device` where ops on subtypes are allowed via certain rules. This leads into a more Explicit view of memory_spaces compared to `Device` (which is more like Auto). Subtyping might make sense for memory_spaces since the rules are very simple and don't create a lot of complexity compared to sharding-in-types.

What would the rules look like?

* Device + Device = Device
* (Device | Hbm | Vmem) + Host = error
* Hbm + Vmem = Hbm | Vmem (based on a rule. For example, `return vmem if aval.dtype.size <= 3MB else hbm`)
* (Device | Hbm | Vmem) + (Device | Hbm | Vmem) = (Device | Hbm | Vmem) based on a rule we can define in the future.

This change **does not** implement the subtyping design described above but we are just noting the fact that it is possible to go into a more granular view of Device on ShapedArray via subtyping and a few simple rules.

Co-authored-by: Matthew Johnson <mattjj@google.com>
PiperOrigin-RevId: 789323712