[jaxpr consts] Disable captured constants warnings with new constant handling
We do not need the captured constants warning when we turn
on the new handling of constants.
See https://docs.jax.dev/en/latest/internals/constants.html.
All constants that are captured will be turned into
arguments, and will not be embedded in the HLO.

Also, drive-by fixes in test_util.assertWarnRegexp (allow
use with `re.compile` patterns, and print a better error message
on failure).