[jaxpr consts] Port captured constants warnings to new handling of constants
Prepare the captured constants warning implementation for when we turn
on the new handling of constants.
See https://docs.jax.dev/en/latest/internals/constants.html.

One difference from the previous implementation is that we now
check for captured constants only once per MLIR module, while
below we did the check for each MLIR function. This is because
we use `core.jaxpr_const_args` which finds the constant in
the whole Jaxpr, and if we did this for every MLIR function
we would end up reporting duplicates. Note that this
changes the behavior in that the warning threshold is now
applied for the entire module. I think this should be Ok.

Also, drive-by fixes in test_util.assertWarnRegexp (allow
use with `re.compile` patterns, and print a better error message
on failure).