[Mosaic GPU] Use a TiledLayout with vector_length=1 for 32-bit WGMMA accumulators
LLVM now treats `<2 x float>` as a native PTX type and uses 64-bit registers to store
them. This, in turn, means that we have to explode them into 32-bit registers right
before WGMMA, which makes ptxas very unhappy and causes it to insert lots of WGMMA waits
that absolutely tank the performance. As a workaround, we use the new WGMMA_LAYOUT_ACC_32BIT
layout when 32-bit data with WGMMA_LAYOUT is used to initialize a WGMMAAccumulator, to ensure
that the LLVM accumulator registers will always be represented as 32-bit PTX registers.

PiperOrigin-RevId: 789716162